Create PolarCircularBase to remove duplicate Code from circulararray and polararray
