+++ src/Mod/Draft/draftguitools/gui_circulararray.py
@@ -25,17 +25,9 @@
 # \ingroup DRAFT
 # \brief This module provides the Draft CircularArray tool.
 
-from pivy import coin
 from PySide.QtCore import QT_TRANSLATE_NOOP
-
-import FreeCAD as App
 import FreeCADGui as Gui
-import Draft
 import Draft_rc  # include resources, icons, ui files
-import draftutils.todo as todo
-
-from draftutils.messages import _msg, _log
-from draftutils.translate import _tr
 from draftguitools import gui_base
 from drafttaskpanels import task_circulararray
 
@@ -43,19 +35,12 @@ from drafttaskpanels import task_circulararray
 bool(Draft_rc.__name__)
 
 
-class CircularArray(gui_base.GuiCommandBase):
+class CircularArray(gui_base.PolarCircularBase):
     """Gui command for the CircularArray tool."""
 
     def __init__(self):
         super(CircularArray, self).__init__()
         self.command_name = "Circular array"
-        self.location = None
-        self.mouse_event = None
-        self.view = None
-        self.callback_move = None
-        self.callback_click = None
-        self.ui = None
-        self.point = App.Vector()
 
     def GetResources(self):
         """Set icon, menu and tooltip."""
@@ -77,68 +62,10 @@ class CircularArray(gui_base.GuiCommandBase):
         We add callbacks that connect the 3D view with
         the widgets of the task panel.
         """
-        _log("GuiCommand: {}".format(_tr(self.command_name)))
-        _msg("{}".format(16*"-"))
-        _msg("GuiCommand: {}".format(_tr(self.command_name)))
-
-        self.location = coin.SoLocation2Event.getClassTypeId()
-        self.mouse_event = coin.SoMouseButtonEvent.getClassTypeId()
-        self.view = Draft.get3DView()
-        self.callback_move = \
-            self.view.addEventCallbackPivy(self.location, self.move)
-        self.callback_click = \
-            self.view.addEventCallbackPivy(self.mouse_event, self.click)
-
         self.ui = task_circulararray.TaskPanelCircularArray()
         # The calling class (this one) is saved in the object
         # of the interface, to be able to call a function from within it.
-        self.ui.source_command = self
-        # Gui.Control.showDialog(self.ui)
-        todo.ToDo.delay(Gui.Control.showDialog, self.ui)
-
-    def move(self, event_cb):
-        """Execute as a callback when the pointer moves in the 3D view.
-
-        It should automatically update the coordinates in the widgets
-        of the task panel.
-        """
-        event = event_cb.getEvent()
-        mousepos = event.getPosition().getValue()
-        ctrl = event.wasCtrlDown()
-        self.point = Gui.Snapper.snap(mousepos, active=ctrl)
-        if self.ui:
-            self.ui.display_point(self.point)
-
-    def click(self, event_cb=None):
-        """Execute as a callback when the pointer clicks on the 3D view.
-
-        It should act as if the Enter key was pressed, or the OK button
-        was pressed in the task panel.
-        """
-        if event_cb:
-            event = event_cb.getEvent()
-            if (event.getState() != coin.SoMouseButtonEvent.DOWN
-                    or event.getButton() != coin.SoMouseButtonEvent.BUTTON1):
-                return
-        if self.ui and self.point:
-            # The accept function of the interface
-            # should call the completed function
-            # of the calling class (this one).
-            self.ui.accept()
-
-    def completed(self):
-        """Execute when the command is terminated.
-
-        We should remove the callbacks that were added to the 3D view
-        and then close the task panel.
-        """
-        self.view.removeEventCallbackPivy(self.location,
-                                          self.callback_move)
-        self.view.removeEventCallbackPivy(self.mouse_event,
-                                          self.callback_click)
-        if Gui.Control.activeDialog():
-            Gui.Control.closeDialog()
-            super(CircularArray, self).finish()
+        super(CircularArray, self).Activated()
 
 
 Gui.addCommand('Draft_CircularArray', CircularArray())
