+++ src/Mod/Draft/draftguitools/gui_polararray.py
@@ -25,18 +25,9 @@
 # \ingroup DRAFT
 # \brief This module provides the Draft PolarArray tool.
 
-from pivy import coin
 from PySide.QtCore import QT_TRANSLATE_NOOP
-
-import FreeCAD as App
 import FreeCADGui as Gui
-import Part
-import Draft
 import Draft_rc  # include resources, icons, ui files
-import draftutils.todo as todo
-
-from draftutils.messages import _msg, _log
-from draftutils.translate import _tr
 from draftguitools import gui_base
 from drafttaskpanels import task_polararray
 
@@ -44,19 +35,12 @@ from drafttaskpanels import task_polararray
 bool(Draft_rc.__name__)
 
 
-class PolarArray(gui_base.GuiCommandBase):
+class PolarArray(gui_base.PolarCircularBase):
     """Gui command for the PolarArray tool."""
 
     def __init__(self):
         super(PolarArray, self).__init__()
         self.command_name = "Polar array"
-        self.location = None
-        self.mouse_event = None
-        self.view = None
-        self.callback_move = None
-        self.callback_click = None
-        self.ui = None
-        self.point = App.Vector()
 
     def GetResources(self):
         """Set icon, menu and tooltip."""
@@ -77,122 +61,12 @@ class PolarArray(gui_base.GuiCommandBase):
 
         We add callbacks that connect the 3D view with
         the widgets of the task panel.
+        super(PolarArray, self).Activated()
         """
-        _log("GuiCommand: {}".format(_tr(self.command_name)))
-        _msg("{}".format(16 * "-"))
-        _msg("GuiCommand: {}".format(_tr(self.command_name)))
-
-        self.location = coin.SoLocation2Event.getClassTypeId()
-        self.mouse_event = coin.SoMouseButtonEvent.getClassTypeId()
-        self.view = Draft.get3DView()
-        self.add_center_callbacks()
-
         self.ui = task_polararray.TaskPanelPolarArray()
         # The calling class (this one) is saved in the object
         # of the interface, to be able to call a function from within it.
-        self.ui.source_command = self
-        # Gui.Control.showDialog(self.ui)
-        todo.ToDo.delay(Gui.Control.showDialog, self.ui)
-
-    def move(self, event_cb):
-        """Execute as a callback when the pointer moves in the 3D view.
-
-        It should automatically update the coordinates in the widgets
-        of the task panel.
-        """
-        event = event_cb.getEvent()
-        mousepos = event.getPosition().getValue()
-        ctrl = event.wasCtrlDown()
-        self.point = Gui.Snapper.snap(mousepos, active=ctrl)
-        if self.ui:
-            self.ui.display_point(self.point)
-
-    def click(self, event_cb=None):
-        """Execute as a callback when the pointer clicks on the 3D view.
-
-        It should act as if the Enter key was pressed, or the OK button
-        was pressed in the task panel.
-        """
-        if event_cb:
-            event = event_cb.getEvent()
-            if (event.getState() != coin.SoMouseButtonEvent.DOWN
-                    or event.getButton() != coin.SoMouseButtonEvent.BUTTON1):
-                return
-        if self.ui and self.point:
-            # The accept function of the interface
-            # should call the completed function
-            # of the calling class (this one).
-            self.ui.accept()
-
-    def add_center_callbacks(self):
-        """Execute when the center selection should be enabled"""
-        self.callback_move = \
-            self.view.addEventCallbackPivy(self.location, self.move)
-        self.callback_click = \
-            self.view.addEventCallbackPivy(self.mouse_event, self.click)
-
-    def remove_center_callbacks(self):
-        """Execute when the center selection should be disabled"""
-        if hasattr(self, "callback_move"):
-            self.view.removeEventCallbackPivy(self.location,
-                                              self.callback_move)
-            del self.callback_move
-
-        if hasattr(self, "callback_click"):
-            self.view.removeEventCallbackPivy(self.mouse_event,
-                                              self.callback_click)
-            del self.callback_click
-
-    def add_axis_selection_observer(self):
-        """Execute when axis reference selection should be enabled"""
-        Gui.Selection.clearSelection(App.ActiveDocument.Name)
-        self.axis_observer = AxisSelectionObserver(self)
-        Gui.Selection.addObserver(self.axis_observer)
-
-    def remove_axis_selection_observer(self):
-        """Execute when axis reference selection should be disabled"""
-        if hasattr(self, "axis_observer"):
-            Gui.Selection.removeObserver(self.axis_observer)
-            del self.axis_observer
-
-    def completed(self):
-        """Execute when the command is terminated.
-
-        We should remove the callbacks that were added to the 3D view
-        and then close the task panel.
-        """
-        self.remove_center_callbacks()
-        self.remove_axis_selection_observer()
-        if Gui.Control.activeDialog():
-            Gui.Control.closeDialog()
-            super(PolarArray, self).finish()
+        super(PolarArray, self).Activated()
 
 
 Gui.addCommand('Draft_PolarArray', PolarArray())
-
-
-class AxisSelectionObserver:
-    """This classes functions will be called when an selection
-    event occurs after axis selection is enabled.
-    """
-
-    def __init__(self, polar_array):
-        self.polar_array = polar_array
-
-    def addSelection(self, doc, obj_name, sub_name, pnt):
-        """Executed when a new selection is added during AxisReference
-        selection process.
-
-        The selection will be cleared and checked if the selected
-        object is edge, if so the axis will be displayed in the UI.
-        """
-        Gui.Selection.clearSelection(App.ActiveDocument.Name)
-        selection = Gui.ActiveDocument.getObject(obj_name)
-        selection_object = selection.Object
-        edge = selection_object.getSubObject(sub_name)
-        if isinstance(edge, Part.Edge) and isinstance(edge.Curve, Part.Line):
-            self.polar_array.ui.display_axis(obj_name, sub_name)
-            self.polar_array.remove_axis_selection_observer()
-        else:
-            self.polar_array.ui.disable_axis_selection()
-            raise TypeError("Selected object is not an edge.")
